fire-alarm/python/requirements.txt
APScheduler==3.2.0
paho-mqtt==1.2
PyEventEmitter==1.0.3
requests==2.11.1
simplejson==3.8.2
twilio==5.6.0